move common eigen calculations in DataFormats/Math
